requirements.txt
langchain>=0.2.0
langchain-community>=0.2.0
langchain-text-splitters>=0.2.0
langchain-anthropic>=0.1.0
anthropic>=0.28.0
faiss-cpu>=1.8.0
pypdf>=4.2.0
streamlit>=1.35.0
python-dotenv>=1.0.1
sentence-transformers>=3.0.0
